refactor: Number로 클래스 분리
